adjust type follow wm type for vd58 to 510i

+++ src/main/java/com/kpei/mkbd/datatransform/dto/Vd510eDto.java
@@ -2,6 +2,8 @@ package com.kpei.mkbd.datatransform.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 @Builder
@@ -9,7 +11,7 @@ import lombok.*;
 @NoArgsConstructor
 public class Vd510eDto extends Vd5Base {
     private String kodeEfek;
-    private double volume;
-    private double harga;
-    private double nilaiPasarWajar;
+    private BigDecimal volume;
+    private BigDecimal harga;
+    private BigDecimal nilaiPasarWajar;
 }
