+++ src/main/java/com/kpei/mkbd/datatransform/dto/Vd59Dto.java
@@ -2,12 +2,14 @@ package com.kpei.mkbd.datatransform.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class Vd59Dto extends Vd5Base {
-    private double jumlah;
-    private double total;
+    private BigDecimal jumlah;
+    private BigDecimal total;
 }
